fix: plug protocol crashes, LPD wiring, type safety, infrastructure smells
Protocol bugs:
- TrackerServer/UdpTrackerServer: replace Split('.') IP parsing with IPAddress.TryParse + GetAddressBytes(); malformed IPs skipped not crash
- TrackerServer: compact peers now use BString(byte[]) not Latin1 string encoding; binary wire format correct
- LocalPeerDiscovery: wire parsed infoHash+port into IPeerDiscoveryService.AddPeers; LPD peer discovery was completely non-functional
- PeerConnection: byte shifts cast to uint before shift then int; add length<0 guard
- PeerServer: same unsigned-shift fix for index/begin/length in piece request handler
- MseHandshake: ReadUint16 returns ushort not int

Infrastructure:
- Startup: Uri.TryCreate for Origin header; UriFormatException no longer crashable by client
- AssemblyLoader: narrow catch to FileNotFoundException/BadImageFormatException/FileLoadException; log at Warn
- BuildInfo: narrow catch to IOException/FormatException; log at Warn
- StringExtensions: remove dead IsNullOrWhiteSpace/IsNotNullOrWhiteSpace extension methods
- RestController: remove wasted validator allocations in base constructor; properties default null

Frontend:
- Statistics: log tracker URL parse failures instead of silently swallowing

Tests:
- LocalPeerDiscoveryTest: mock IPeerDiscoveryService; assert AddPeers called with correct infoHash+port
- PeerConnectionTest: new test for length<0 rejection on high-bit-set length bytes
- PeerServerTest: 3 new tests for high-bit-set index/begin/length rejection

+++ src/NzbDrone.Core/Peers/PeerServer.cs
@@ -470,9 +470,9 @@ public class PeerServer : BackgroundService
 
     private static void HandlePieceRequest(PeerConnection connection, byte[] payload)
     {
-        var index = (payload[0] << 24) | (payload[1] << 16) | (payload[2] << 8) | payload[3];
-        var begin = (payload[4] << 24) | (payload[5] << 16) | (payload[6] << 8) | payload[7];
-        var length = (payload[8] << 24) | (payload[9] << 16) | (payload[10] << 8) | payload[11];
+        var index = (int)(((uint)payload[0] << 24) | ((uint)payload[1] << 16) | ((uint)payload[2] << 8) | payload[3]);
+        var begin = (int)(((uint)payload[4] << 24) | ((uint)payload[5] << 16) | ((uint)payload[6] << 8) | payload[7]);
+        var length = (int)(((uint)payload[8] << 24) | ((uint)payload[9] << 16) | ((uint)payload[10] << 8) | payload[11]);
 
         const int MaxBlockSize = 32768; // 2x standard 16KB block size
         if (length <= 0 || length > MaxBlockSize)
